GridFieldOrderableRows: Support extra sort fields
This change allows for the specification of extra sort fields that are to be
applied before the user-defined sort field's order. This is useful in
scenarios where a GridField is used with one or more "fixed" columns which
eg. categorize the entries and whose order is important but the user should
still be allowed to modify the order within a category.

This change affects the PHP code only, therefore the Javascript code will
still let the user drag rows into foreign "categories". However, after
dropping it the GridField will reload and the extra sort order will be
enforced again.