GridFieldOrderableRows: Allow setting the reorder handles column's number
Until now the column with the handles starting the reorder action always
appeared as the first column at the left of the GridField. As a followup to
the previous change's for extra sort fields, this change allows specifying
a different column index. This way you can eg. move the handles behind a
category column whose sorting will always be enforced, thus giving the user
visual support in understanding the limits of drag-n-drop reordering.