Add a new component for a pop-out 'meatball' menu
Which will allow per item (row) specific actions
such as grouping things like jumping to a specific tab
in the DetailForm, and if the displayed model is Versioned
then showing relevant (contextual) actions for Publishing,
Unpublishing, and Archiving (deleting) an item. This saves
a user from having to navigate through depths for each
record they wish to perform an action on, without crowding
the actions column for a row on the GridField.

Also included in this commit source structural updates
with regards to updating to a modern front end build
system consistent with the silverstripe/admin core.
This allows the module the advantage of being able to
include CMS centric components without having a hard
dependency to the project. This should cover a good
90%+ use cases, but allow for flexibility for framework
only builds wishing to use this module.