Update ActionsMenu component to be more extensible
In order to support other use cases for this menu, such as developers
adding their own items or sections to various other actions which may be
allowed on their items, we can make use of constructing a default state
which can then be modified, rather than building out an unchangable state
when rendering.

To this end what was previously tightly coupled to other
classes such as TabSet are now decoupled via the use of closures in the
menu construction. A developer now has the ability to either build a
static array of items, or use a closure function to return such an array -
one per group. Furthermore empty items and sections are now cleaned before
output.