Alter RecordActionHandler's default to be more extensible
Previously we've had the $Action-less request mimic the default behaviour
of GridFieldDetailFrom, that is to call that class' "edit" action. However
in future a developer may wish to alter that behaviour, and formerly it
was infeasible to do so without great trickery. So this commit introduces
a configuration setting to allow the default action to be set as such, of
which the default is the former (previously described) behaviour. Now a
developer can apply an `Extension` class and not only add new actions that
can be called, but now allow one of those added actions to become the
default call when no action is supplied in the request.