NEW Control which components are written on GridFieldEditableColumns::handleSave
Many objects contain a link back to the parent object (by nature of
has_many), the same object that is being saved and triggering the
GridFieldComponent to save its relational components in turn. This is
often unnecessary and creates a 2N scenario for each record in the list,
affecting performance. Especially when other onBefore/After/During write
hooks are hanging off that class (e.g. fulltext search updates).